# openapi/notes-api-v1.yaml
...
openapi: 3.0.0
info:
  title: OSM Notes API
  version: '0.1.0'
  description: |
    REST API for OSM Notes Analytics and Ingestion. Provides programmatic access to
    user profiles, country analytics, notes, and real-time metrics.

    Responses that contain data from OpenStreetMap or derivatives include an
    `attribution` field (text and url) so clients can display proper attribution
    (e.g. "© OpenStreetMap contributors").
  contact:
    name: OSM Notes Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://notes-api.osm.lat
    description: Production server
security:
  - UserAgent: []
components:
  securitySchemes:
    UserAgent:
      type: apiKey
      in: header
      name: User-Agent
      description: 'User-Agent header required for all requests. Format: AppName/Version (Contact)'
  schemas:
    Error:
      type: object
      required: [error, message, statusCode]
      properties:
        error:
          type: string
          description: Error type name
          example: Bad Request
        message:
          type: string
          description: Human-readable error message
          example: Invalid note ID
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
...
    UserProfile:
      type: object
      properties:
        dimension_user_id:
          type: integer
          example: 123
        user_id:
          type: integer
          example: 12345
        username:
          type: string
          nullable: true
          example: test_user
        history_whole_open:
          type: integer
          example: 100
        history_whole_closed:
          type: integer
          example: 50
        history_whole_commented:
          type: integer
          example: 75
        avg_days_to_resolution:
          type: number
          format: float
          nullable: true
          example: 5.5
        resolution_rate:
          type: number
          format: float
          nullable: true
          example: 50.0
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 250
        total_pages:
          type: integer
          example: 13
...
paths:
...
  /notes-api/v1/users:
    get:
      summary: List all users with pagination
      tags: [Users]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
        - in: query
          name: sort
          schema:
            type: string
            enum: [user_id, username, history_whole_open, history_whole_closed, resolution_rate]
            default: user_id
          description: Sort field
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Paginated list of users
          headers:
            X-Total-Count:
              description: Total number of users
              schema:
                type: integer
            X-Page:
              description: Current page number
              schema:
                type: integer
            X-Per-Page:
              description: Items per page
              schema:
                type: integer
            X-Total-Pages:
              description: Total number of pages
              schema:
                type: integer
            Link:
              description: Pagination links (RFC 5988)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserProfile'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/users/{user_id}:
    get:
      summary: Get user profile by ID
      tags: [Users]
      security:
        - UserAgent: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: OSM user ID
      responses:
        '200':
          description: User profile with analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notes-api/v1/search/users:
    get:
      summary: Search users by username or user_id
      tags: [Search]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query (username pattern or user_id)
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: integer
                        username:
                          type: string
                        history_whole_open:
                          type: integer
                        history_whole_closed:
                          type: integer
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes-api/v1/search/countries:
    get:
      summary: Search countries by name, ISO code, or country_id
      tags: [Search]
      security:
        - UserAgent: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Search query (name pattern, ISO code, or country_id)
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        country_id:
                          type: integer
                        country_name:
                          type: string
                        iso_alpha2:
                          type: string
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...